+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog of threedigrid-builder
 1.5.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Use the global 'max_infiltration_capacity', if present.
 
 
 1.5.0 (2022-10-26)